# tests/contract/api/openapi.json
{
  "components": {
    "schemas": {
      "AnalysisAccepted": {
        "additionalProperties": false,
        "description": "Asynchronous analysis creation. Poll status; do not assume completion.",
        "properties": {
          "analysis_id": {
            "format": "uuid",
            "title": "Analysis Id",
            "type": "string"
          },
          "configuration_hash": {
            "title": "Configuration Hash",
            "type": "string"
          },
          "pipeline_version": {
            "title": "Pipeline Version",
            "type": "string"
          },
          "reused": {
            "title": "Reused",
            "type": "boolean"
          },
          "state": {
            "$ref": "#/components/schemas/AnalysisState"
          },
          "video_id": {
            "format": "uuid",
            "title": "Video Id",
            "type": "string"
          }
        },
        "required": [
          "analysis_id",
          "video_id",
          "configuration_hash",
          "pipeline_version",
          "reused",
          "state"
        ],
        "title": "AnalysisAccepted",
        "type": "object"
      },
...
      "AnalysisCreateRequest": {
        "additionalProperties": false,
        "description": "Start or reuse an analysis for an existing video.",
        "properties": {
          "config": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AnalysisConfig"
              },
              {
                "type": "null"
              }
            ]
          },
          "video_id": {
            "format": "uuid",
            "title": "Video Id",
            "type": "string"
          }
        },
        "required": [
          "video_id"
        ],
        "title": "AnalysisCreateRequest",
        "type": "object"
      },
      "AnalysisReport": {
        "additionalProperties": false,
        "description": "The versioned document a client reads. Large series live behind artifact references.",
        "properties": {
          "analysis_id": {
            "format": "uuid",
            "title": "Analysis Id",
            "type": "string"
          },
          "availability": {
            "$ref": "#/components/schemas/ReportAvailability"
          },
          "configuration_hash": {
            "description": "Lowercase hex SHA-256 digest.",
            "pattern": "^[0-9a-f]{64}$",
            "title": "Configuration Hash",
            "type": "string"
          },
          "critique": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Critique"
              },
              {
                "type": "null"
              }
            ]
          },
          "generated_at": {
            "format": "date-time",
            "title": "Generated At",
            "type": "string"
          },
          "pipeline_version": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Pipeline Version",
            "type": "string"
          },
          "schema_version": {
            "description": "Report schema version.",
            "title": "Schema Version",
            "type": "string"
          },
          "shots": {
            "items": {
              "$ref": "#/components/schemas/ShotAnalysis"
            },
            "minItems": 1,
            "title": "Shots",
            "type": "array"
          },
          "summary": {
            "$ref": "#/components/schemas/VideoSummary"
          },
          "timeline_artifact": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ArtifactRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "video": {
            "$ref": "#/components/schemas/VideoMetadata"
          }
        },
        "required": [
          "schema_version",
          "analysis_id",
          "video",
          "generated_at",
          "pipeline_version",
          "configuration_hash",
          "availability",
          "summary",
          "shots"
        ],
        "title": "AnalysisReport",
        "type": "object"
      },
      "AnalysisState": {
        "description": "Overall analysis state. Progress is derived from stages, not from log text.",
        "enum": [
          "QUEUED",
          "RUNNING",
          "CANCEL_REQUESTED",
          "CANCELED",
          "SUCCEEDED",
          "PARTIAL",
          "FAILED"
        ],
        "title": "AnalysisState",
        "type": "string"
      },
...
      "Body_upload_video_v1_videos_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_upload_video_v1_videos_post",
        "type": "object"
      },
...
      "Critique": {
        "additionalProperties": false,
        "description": "Optional interpretation of an already-validated report. Regenerable; cannot write metrics.",
        "properties": {
          "input_report_sha256": {
            "anyOf": [
              {
                "description": "Lowercase hex SHA-256 digest.",
                "pattern": "^[0-9a-f]{64}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input Report Sha256"
          },
          "model_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Name"
          },
          "prompt_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt Version"
          },
          "status": {
            "$ref": "#/components/schemas/MetricStatus"
          },
          "text": {
            "anyOf": [
              {
                "maxLength": 1200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text"
          }
        },
        "required": [
          "status"
        ],
        "title": "Critique",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ShotAnalysis": {
        "additionalProperties": false,
        "description": "One shot and the three per-shot measurement envelopes.",
        "properties": {
          "chromatic": {
            "$ref": "#/components/schemas/Measurement_ChromaticValue_"
          },
          "shot": {
            "$ref": "#/components/schemas/Shot"
          },
          "spatial": {
            "$ref": "#/components/schemas/Measurement_SpatialValue_"
          },
          "temporal": {
            "$ref": "#/components/schemas/Measurement_TemporalValue_"
          }
        },
        "required": [
          "shot",
          "chromatic",
          "spatial",
          "temporal"
        ],
        "title": "ShotAnalysis",
        "type": "object"
      },
...
      "TimelineWindowResponse": {
        "additionalProperties": false,
        "description": "Bounded timeline slice. Points are already downsampled.",
        "properties": {
          "analysis_id": {
            "format": "uuid",
            "title": "Analysis Id",
            "type": "string"
          },
          "end_ms": {
            "title": "End Ms",
            "type": "integer"
          },
          "max_points": {
            "title": "Max Points",
            "type": "integer"
          },
          "points": {
            "items": {
              "$ref": "#/components/schemas/TimelinePoint"
            },
            "title": "Points",
            "type": "array"
          },
          "start_ms": {
            "title": "Start Ms",
            "type": "integer"
          }
        },
        "required": [
          "analysis_id",
          "start_ms",
          "end_ms",
          "max_points",
          "points"
        ],
        "title": "TimelineWindowResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "VideoAccepted": {
        "additionalProperties": false,
        "description": "Identity returned after a streamed upload.",
        "properties": {
          "content_sha256": {
            "title": "Content Sha256",
            "type": "string"
          },
          "duration_ms": {
            "title": "Duration Ms",
            "type": "integer"
          },
          "has_audio": {
            "title": "Has Audio",
            "type": "boolean"
          },
          "height": {
            "title": "Height",
            "type": "integer"
          },
          "reused": {
            "title": "Reused",
            "type": "boolean"
          },
          "video_id": {
            "format": "uuid",
            "title": "Video Id",
            "type": "string"
          },
          "width": {
            "title": "Width",
            "type": "integer"
          }
        },
        "required": [
          "video_id",
          "content_sha256",
          "reused",
          "duration_ms",
          "width",
          "height",
          "has_audio"
        ],
        "title": "VideoAccepted",
        "type": "object"
      },
...
    }
  },
  "info": {
    "summary": "Reproducible cinematography measurement control plane.",
    "title": "cine-analyzer",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/analyses": {
      "post": {
        "description": "Queue an analysis. Video identity stays separate from config identity.",
        "operationId": "create_analysis_v1_analyses_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalysisCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Analysis",
        "tags": [
          "analyses"
        ]
      }
    },
    "/v1/analyses/{analysis_id}": {
      "get": {
        "description": "Status, progress, and unavailable pillars.",
        "operationId": "get_analysis_v1_analyses__analysis_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "analysis_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Analysis Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Analysis",
        "tags": [
          "analyses"
        ]
      }
    },
    "/v1/analyses/{analysis_id}/cancel": {
      "post": {
        "description": "Request cooperative cancellation.",
        "operationId": "cancel_analysis_v1_analyses__analysis_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "analysis_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Analysis Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Analysis",
        "tags": [
          "analyses"
        ]
      }
    },
    "/v1/analyses/{analysis_id}/critique": {
      "get": {
        "description": "Stored interpretation. Omitted when the critic did not run.",
        "operationId": "get_critique_v1_analyses__analysis_id__critique_get",
        "parameters": [
          {
            "in": "path",
            "name": "analysis_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Analysis Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Critique"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Critique",
        "tags": [
          "analyses"
        ]
      }
    },
    "/v1/analyses/{analysis_id}/report": {
      "get": {
        "description": "Canonical AnalysisReport JSON once aggregation has finished.",
        "operationId": "get_report_v1_analyses__analysis_id__report_get",
        "parameters": [
          {
            "in": "path",
            "name": "analysis_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Analysis Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisReport"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Report",
        "tags": [
          "analyses"
        ]
      }
    },
    "/v1/analyses/{analysis_id}/shots/{shot_id}": {
      "get": {
        "description": "One shot's measured metrics from the canonical report.",
        "operationId": "get_shot_v1_analyses__analysis_id__shots__shot_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "analysis_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Analysis Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "shot_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Shot Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShotAnalysis"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Shot",
        "tags": [
          "analyses"
        ]
      }
    },
    "/v1/analyses/{analysis_id}/timeline": {
      "get": {
        "description": "Windowed, downsampled tension-proxy timeline.",
        "operationId": "get_timeline_v1_analyses__analysis_id__timeline_get",
        "parameters": [
          {
            "in": "path",
            "name": "analysis_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Analysis Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "start_ms",
            "required": true,
            "schema": {
              "minimum": 0,
              "title": "Start Ms",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "end_ms",
            "required": true,
            "schema": {
              "minimum": 0,
              "title": "End Ms",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "max_points",
            "required": true,
            "schema": {
              "maximum": 2000,
              "minimum": 1,
              "title": "Max Points",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimelineWindowResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Timeline",
        "tags": [
          "analyses"
        ]
      }
    },
    "/v1/artifacts/{artifact_id}": {
      "get": {
        "description": "Authorized stream. Possession of the server-issued UUID is the capability.",
        "operationId": "get_artifact_v1_artifacts__artifact_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "artifact_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Artifact Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Artifact",
        "tags": [
          "artifacts"
        ]
      }
    },
    "/v1/videos": {
      "post": {
        "description": "Stream an upload into the Phase 03 ingest use case.",
        "operationId": "upload_video_v1_videos_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_video_v1_videos_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Video",
        "tags": [
          "videos"
        ]
      }
    }
  }
}